Inline completion file renames (#226419)
