+++ src/vs/editor/contrib/inlineCompletions/browser/model/ghostText.ts
