+++ src/vs/editor/contrib/inlineCompletions/browser/view/ghostTextView.css
