+++ src/vs/editor/contrib/inlineCompletions/browser/view/ghostTextView.ts
@@ -8,7 +8,7 @@ import { Event } from 'vs/base/common/event';
 import { Disposable, toDisposable } from 'vs/base/common/lifecycle';
 import { IObservable, autorun, derived, observableFromEvent, observableSignalFromEvent, observableValue } from 'vs/base/common/observable';
 import * as strings from 'vs/base/common/strings';
-import 'vs/css!./ghostText';
+import 'vs/css!./ghostTextView';
 import { applyFontInfo } from 'vs/editor/browser/config/domFontInfo';
 import { ICodeEditor } from 'vs/editor/browser/editorBrowser';
 import { EditorFontLigatures, EditorOption, IComputedEditorOptions } from 'vs/editor/common/config/editorOptions';
@@ -22,7 +22,7 @@ import { LineTokens } from 'vs/editor/common/tokens/lineTokens';
 import { LineDecoration } from 'vs/editor/common/viewLayout/lineDecorations';
 import { RenderLineInput, renderViewLine } from 'vs/editor/common/viewLayout/viewLineRenderer';
 import { InlineDecorationType } from 'vs/editor/common/viewModel';
-import { GhostText, GhostTextReplacement } from 'vs/editor/contrib/inlineCompletions/browser/ghostText';
+import { GhostText, GhostTextReplacement } from 'vs/editor/contrib/inlineCompletions/browser/model/ghostText';
 import { ColumnRange, applyObservableDecorations } from 'vs/editor/contrib/inlineCompletions/browser/utils';
 
 export const GHOST_TEXT_DESCRIPTION = 'ghost-text';
@@ -32,7 +32,7 @@ export interface IGhostTextWidgetModel {
 	readonly minReservedLineCount: IObservable<number>;
 }
 
-export class GhostTextWidget extends Disposable {
+export class GhostTextView extends Disposable {
 	private readonly isDisposed = observableValue(this, false);
 	private readonly currentTextModel = observableFromEvent(this, this.editor.onDidChangeModel, () => /** @description editor.model */ this.editor.getModel());
