+++ src/vs/editor/contrib/inlineCompletions/test/browser/inlineCompletionsModel.test.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 import assert from 'assert';
 import { Position } from 'vs/editor/common/core/position';
-import { getSecondaryEdits } from 'vs/editor/contrib/inlineCompletions/browser/inlineCompletionsModel';
+import { getSecondaryEdits } from 'vs/editor/contrib/inlineCompletions/browser/model/inlineCompletionsModel';
 import { SingleTextEdit } from 'vs/editor/common/core/textEdit';
 import { createTextModel } from 'vs/editor/test/common/testTextModel';
 import { Range } from 'vs/editor/common/core/range';
